// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SALES_EXEC: 'SALES_EXEC',
  SALES_MANAGER: 'SALES_MANAGER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProposalStatus = {
  DRAFT: 'DRAFT',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  REVISION_REQUIRED: 'REVISION_REQUIRED',
  APPROVED: 'APPROVED',
  SENT: 'SENT',
  WON: 'WON',
  LOST: 'LOST',
  ON_HOLD: 'ON_HOLD',
  EXPIRED: 'EXPIRED'
} as const

export type ProposalStatus = (typeof ProposalStatus)[keyof typeof ProposalStatus]